Revert "Release/arthur"
