+++ src/app/home/[subjectID]/setup-quiz/quiz/page.tsx
@@ -3,7 +3,9 @@ import React, { useMemo, useEffect, useState } from 'react';
 import { useParams, useRouter } from "next/navigation";
 import { Question } from '../../../../../types/api/Question';
 import ProtectedPage from '../../../../../components/ProtectPage';
+import { Bookmark, BookmarkBorder, CheckCircle, Cancel, ErrorOutline, ViewList, ViewModule, Visibility, VisibilityOff } from '@mui/icons-material';
 import { useUser } from '../../../../../hooks/User/useUser';
+import ImageGallery from '../../../../../components/magicui/ImageGallery';
 import { useGetQuizzes } from '../../../../../hooks/quiz/useGetQuizzes';
 import { useSubmitScore } from '../../../../../hooks/score/useSubmitScore';
 import { Quiz } from '../../../../../types/api/Quiz';
@@ -11,18 +13,6 @@ import { useGetUserStatById } from '../../../../../hooks/stats/useGetUserStatByI
 import { Role_type } from '../../../../../config/role';
 import { useQuiz } from '../../../../../context/quiz'
 import { FrontendRoutes } from '../../../../../config/apiRoutes'
-import { useShuffledChoices } from '../../../../../hooks/quiz/useShuffledChoices';
-import { useQuizSessionStorage } from '../../../../../hooks/quiz/useQuizSessionStorage';
-import { useQuizKeyboardNavigation } from '../../../../../hooks/quiz/useQuizKeyboardNavigation';
-import QuizHeader from '../../../../../components/quiz/quizQuestion/QuizHeader';
-import QuizNavigation from '../../../../../components/quiz/quizQuestion/QuizNavigation';
-import QuizQuestion from '../../../../../components/quiz/quizQuestion/QuizQuestion';
-import QuizActions from '../../../../../components/quiz/quizQuestion/QuizActions';
-import CorrectAnswerDisplay from '../../../../../components/quiz/quizQuestion/CorrectAnswerDisplay';
-import QuizSubmitButtons from '../../../../../components/quiz/quizQuestion/QuizSubmitButtons';
-import QuizQuestionTable from '../../../../../components/quiz/quizQuestion/QuizQuestionTable';
-import QuizImageModal from '../../../../../components/quiz/quizQuestion/QuizImageModal';
-import KeyboardShortcutsHelp from '../../../../../components/quiz/KeyboardShortcutsHelp';
 
 
 export default function Problem() {
@@ -35,13 +25,11 @@ export default function Problem() {
     const isSAdmin = user?.role === Role_type.SADMIN;
     const isAdmin = user?.role === Role_type.ADMIN || isSAdmin;
     const subjectID = params.subjectID as string;
-    
     useEffect(() => {
         if (!selectCategory.length || questionCount <= 0 || !answerMode || !selectedQuestionTypes) {
-            router.replace(`${FrontendRoutes.HOMEPAGE}/${subjectID}/setup-quiz`)
+            router.replace(`${FrontendRoutes.HOMEPAGE}/${subjectID}/quiz`)
         }
     }, [selectCategory, questionCount, answerMode, selectedQuestionTypes, router, subjectID]);
-    
     const { data: userStat, isLoading: statLoading } = useGetUserStatById(user?._id || '', subjectID, !!user?._id && !!subjectID);
     const canTakeQuiz = isAdmin || (userStat?.quizCount ?? 0) >= 4;
 
@@ -49,18 +37,11 @@ export default function Problem() {
     const [showQuestion, setShowQuestion] = useState<Question[]>([]);
     const [isImageModalOpen, setIsImageModalOpen] = useState(false);
     const [currentQuestionIndex, setCurrentQuestionIndex] = useState(0);
+    const [currentImageIndex, setCurrentImageIndex] = useState(0);
     const [isQuestionTableOpen, setIsQuestionTableOpen] = useState(false);
     const [questionViewMode, setQuestionViewMode] = useState<'grid' | 'list'>('grid');
     const [isSubmitting, setIsSubmitting] = useState(false);
     const [isTimerVisible, setIsTimerVisible] = useState(true);
-
-    // Use custom hooks
-    const { clearQuizSessionStorage } = useQuizSessionStorage({
-        subjectID,
-        selectedQuestionTypes: selectedQuestionTypes || '',
-        questionCount,
-        selectCategory
-    });
 
     const { data: quizData, isLoading } = useGetQuizzes({
         subjectID,
@@ -86,29 +67,6 @@ export default function Problem() {
                 return newArray;
             };
 
-            // Create a unique session key based on quiz parameters
-            const sessionKey = `quiz_${subjectID}_${selectedQuestionTypes}_${questionCount}_${selectCategory.sort().join('_')}`;
-            
-            // Check if shuffled questions exist in session storage
-            const cachedQuestions = sessionStorage.getItem(sessionKey);
-            if (cachedQuestions) {
-                try {
-                    const parsedQuestions = JSON.parse(cachedQuestions);
-                    // Validate that the cached questions are still valid
-                    const validQuestions = parsedQuestions.filter((q: Question) => 
-                        quizzes.some(quiz => quiz._id === q.quiz._id)
-                    );
-                    
-                    if (validQuestions.length === parsedQuestions.length) {
-                        return {
-                            questions: validQuestions,
-                        };
-                    }
-                } catch (error) {
-                    console.warn('Failed to parse cached questions:', error);
-                }
-            }
-
             const allQuestions = mapToQuestion(quizzes);
 
             const filteredQuestions = allQuestions.filter((q) => {
@@ -122,13 +80,6 @@ export default function Problem() {
             const shuffledQuestions = shuffleArray(filteredQuestions);
             const limitedQuestions = shuffledQuestions.slice(0, questionCount);
 
-            // Save shuffled questions to session storage
-            try {
-                sessionStorage.setItem(sessionKey, JSON.stringify(limitedQuestions));
-            } catch (error) {
-                console.warn('Failed to save questions to session storage:', error);
-            }
-
             return {
                 questions: limitedQuestions,
             };
@@ -137,96 +88,11 @@ export default function Problem() {
 
     const { submitScore, calculateScore } = useSubmitScore();
 
-    const currentQuestion = showQuestion[currentQuestionIndex];
-    
-    const shuffledChoices = useShuffledChoices({
-        currentQuestion,
-        selectedQuestionTypes: selectedQuestionTypes || '',
-        subjectID,
-        questionCount,
-        selectCategory
-    });
-
     useEffect(() => {
         if (quizData?.questions) {
-            // Check if we have saved progress in session storage
-            const sessionKey = `quiz_${subjectID}_${selectedQuestionTypes}_${questionCount}_${selectCategory.sort().join('_')}`;
-            const savedProgress = sessionStorage.getItem(`${sessionKey}_progress`);
-            
-            if (savedProgress) {
-                try {
-                    const parsedProgress = JSON.parse(savedProgress);
-                    // Merge saved answers with fresh questions
-                    const questionsWithProgress = quizData.questions.map((question, index) => {
-                        const savedQuestion = parsedProgress[index];
-                        if (savedQuestion && savedQuestion.quiz._id === question.quiz._id) {
-                            return {
-                                ...question,
-                                select: savedQuestion.select,
-                                isBookmarked: savedQuestion.isBookmarked,
-                                isAnswered: savedQuestion.isAnswered,
-                                isSubmitted: savedQuestion.isSubmitted,
-                                isCorrect: savedQuestion.isCorrect
-                            };
-                        }
-                        return question;
-                    });
-                    setShowQuestion(questionsWithProgress);
-                } catch (error) {
-                    console.warn('Failed to parse saved progress:', error);
-                    setShowQuestion(quizData.questions);
-                }
-            } else {
-                setShowQuestion(quizData.questions);
-            }
+            setShowQuestion(quizData.questions);
         }
-    }, [quizData, subjectID, selectedQuestionTypes, questionCount, selectCategory]);
-
-    useEffect(() => {
-        if (showQuestion.length > 0) {
-            const sessionKey = `quiz_${subjectID}_${selectedQuestionTypes}_${questionCount}_${selectCategory.sort().join('_')}`;
-            
-            const timeoutId = setTimeout(() => {
-                try {
-                    sessionStorage.setItem(`${sessionKey}_progress`, JSON.stringify(showQuestion));
-                } catch (error) {
-                    console.warn('Failed to save progress to session storage:', error);
-                }
-            }, 500);
-
-            return () => clearTimeout(timeoutId);
-        }
-    }, [showQuestion, subjectID, selectedQuestionTypes, questionCount, selectCategory]);
-
-    useEffect(() => {
-        const sessionKey = `quiz_${subjectID}_${selectedQuestionTypes}_${questionCount}_${selectCategory.sort().join('_')}`;
-        
-        const saveProgress = () => {
-            if (showQuestion.length > 0) {
-                try {
-                    sessionStorage.setItem(`${sessionKey}_progress`, JSON.stringify(showQuestion));
-                } catch (error) {
-                    console.warn('Failed to save progress to session storage:', error);
-                }
-            }
-        };
-
-        return () => {
-            saveProgress();
-        };
-    }, [showQuestion, subjectID, selectedQuestionTypes, questionCount, selectCategory]);
-
-    // Save progress when navigating between questions
-    useEffect(() => {
-        if (showQuestion.length > 0) {
-            const sessionKey = `quiz_${subjectID}_${selectedQuestionTypes}_${questionCount}_${selectCategory.sort().join('_')}`;
-            try {
-                sessionStorage.setItem(`${sessionKey}_progress`, JSON.stringify(showQuestion));
-            } catch (error) {
-                console.warn('Failed to save progress to session storage:', error);
-            }
-        }
-    }, [currentQuestionIndex]); // Only save when question index changes
+    }, [quizData]);
 
     useEffect(() => {
         const interval = setInterval(() => {
@@ -236,11 +102,21 @@ export default function Problem() {
         return () => clearInterval(interval);
     }, []);
 
+    const formatTime = (s: number) => {
+        const mins = Math.floor(s / 60);
+        const secs = s % 60;
+        return `${String(mins).padStart(2, '0')}:${String(secs).padStart(2, '0')}`;
+    };
+
+    const currentQuestion = showQuestion[currentQuestionIndex];
+
     const handleQuestionNavigation = (direction: 'next' | 'previous') => {
         if (direction === 'next') {
             setCurrentQuestionIndex((prevIndex: number) => (prevIndex + 1) % showQuestion.length);
+            setCurrentImageIndex(0);
         } else if (direction === 'previous') {
             setCurrentQuestionIndex((prevIndex: number) => (prevIndex - 1 + showQuestion.length) % showQuestion.length);
+            setCurrentImageIndex(0);
         }
     };
 
@@ -294,7 +170,6 @@ export default function Problem() {
         try {
             const result = await submitScore(scoreData);
             if (result?._id) {
-                clearQuizSessionStorage();
                 router.push(`/profile/${result._id}`);
             } else {
                 console.error('No score ID returned from submission');
@@ -379,17 +254,19 @@ export default function Problem() {
         Array.isArray(showQuestion) && showQuestion.length > 0 && showQuestion.every(q => q.isSubmitted)
     , [showQuestion]);
 
-    // Add keyboard navigation
-    useQuizKeyboardNavigation({
-        currentQuestion,
-        currentQuestionIndex,
-        totalQuestions: showQuestion.length,
-        selectedQuestionTypes: selectedQuestionTypes || '',
-        shuffledChoices,
-        onNavigateQuestion: handleQuestionNavigation,
-        onAnswerSelection: handleAnswerSelection,
-        isSubmitted: currentQuestion?.isSubmitted || false
-    });
+    // Add shuffled choices memo
+    const shuffledChoices = useMemo(() => {
+        if (!currentQuestion || selectedQuestionTypes !== 'mcq') return [];
+        
+        // Create a copy of choices with their indices
+        const choicesWithIndices = currentQuestion.quiz.choice.map((choice, index) => ({
+            choice,
+            originalIndex: index
+        }));
+        
+        // Shuffle the choices
+        return [...choicesWithIndices].sort(() => Math.random() - 0.5);
+    }, [currentQuestion, selectedQuestionTypes]);
 
     if (userLoading || statLoading) {
         return (
@@ -425,76 +302,276 @@ export default function Problem() {
             <div className="min-h-screen bg-gray-50 py-8">
                 <div className="container mx-auto px-4 sm:px-6 lg:px-8 max-w-7xl">
                     <div className="bg-white rounded-2xl shadow-xl p-6 sm:p-8">
-                        <QuizHeader
-                            subjectName={showQuestion[0]?.quiz?.subject?.name || ''}
-                            currentQuestionIndex={currentQuestionIndex}
-                            totalQuestions={showQuestion.length}
-                            seconds={seconds}
-                            isTimerVisible={isTimerVisible}
-                            onToggleTimer={() => setIsTimerVisible((v) => !v)}
-                        />
-
-                        <QuizNavigation
-                            currentQuestionIndex={currentQuestionIndex}
-                            totalQuestions={showQuestion.length}
-                            onNavigate={handleQuestionNavigation}
-                        />
-
-                        <div className="flex flex-col items-center justify-center mb-8 sm:mb-12">
-                            {currentQuestion && (
-                                <>
-                                    <QuizQuestion
-                                        currentQuestion={currentQuestion}
-                                        currentQuestionIndex={currentQuestionIndex}
-                                        selectedQuestionTypes={selectedQuestionTypes || ''}
-                                        shuffledChoices={shuffledChoices}
-                                        onBookmarkToggle={toggleBookmark}
-                                        onAnswerSelection={handleAnswerSelection}
-                                        onShortAnswerChange={handleShortAnswerChange}
-                                    />
-                                    
-                                    <QuizActions
-                                        currentQuestion={currentQuestion}
-                                        selectedQuestionTypes={selectedQuestionTypes || ''}
-                                        answerMode={answerMode || ''}
-                                        onSubmitCurrentQuestion={submitCurrentQuestion}
-                                        onClearAnswer={clearAnswer}
-                                    />
-                                    
-                                    <CorrectAnswerDisplay currentQuestion={currentQuestion} />
-                                </>
+                <div className="text-center mb-8">
+                            <h1 className="text-3xl sm:text-4xl font-extrabold mb-4 sm:mb-6 bg-gradient-to-r from-blue-600 to-blue-800 bg-clip-text text-transparent">
+                                {showQuestion[0]?.quiz?.subject?.name || 'Loading...'}
+                            </h1>
+                    <div className="flex justify-between items-center">
+                        <p className="text-base sm:text-lg text-gray-600 mt-2 sm:mt-4 font-medium">
+                            Question {currentQuestionIndex + 1} of {showQuestion.length}
+                        </p>
+                        <div className="flex items-center gap-2">
+                            {isTimerVisible && (
+                                <div className="bg-blue-100 text-blue-800 px-4 py-2 rounded-lg font-semibold shadow-sm">
+                                    Time: {formatTime(seconds)}
+                                </div>
                             )}
+                            <button
+                                className="ml-2 px-2 py-1 text-xs bg-gray-200 hover:bg-gray-300 rounded transition flex items-center justify-center"
+                                onClick={() => setIsTimerVisible((v) => !v)}
+                                title={isTimerVisible ? 'Hide Timer' : 'Show Timer'}
+                            >
+                                {isTimerVisible ? <VisibilityOff /> : <Visibility />}
+                            </button>
                         </div>
-
-                        <QuizSubmitButtons
-                            answerMode={answerMode || ''}
-                            allQuestionsAnswered={allQuestionsAnswered}
-                            allQuestionsSubmitted={allQuestionsSubmitted}
-                            isSubmitting={isSubmitting}
-                            onSubmit={handleSubmit}
-                            onViewAllQuestions={() => setIsQuestionTableOpen(true)}
-                        />
                     </div>
                 </div>
-            </div>
-            
-            <QuizQuestionTable
-                isOpen={isQuestionTableOpen}
-                showQuestion={showQuestion}
-                currentQuestionIndex={currentQuestionIndex}
-                questionViewMode={questionViewMode}
-                onClose={() => setIsQuestionTableOpen(false)}
-                onToggleViewMode={toggleQuestionViewMode}
-                onNavigateToQuestion={navigateToQuestion}
-            />
-            
-            <QuizImageModal
-                isOpen={isImageModalOpen}
-                currentQuestion={currentQuestion}
-                onClose={() => setIsImageModalOpen(false)}
-            />
-            
-            <KeyboardShortcutsHelp />
+
+                        <div className="flex justify-between items-center gap-3 sm:gap-4 mb-8">
+                    <button
+                                className="px-4 sm:px-6 py-2.5 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-all duration-300 text-sm sm:text-base font-medium shadow-md hover:shadow-lg transform hover:-translate-y-0.5 disabled:opacity-50 disabled:cursor-not-allowed"
+                        onClick={(e) => {
+                            e.stopPropagation();
+                            handleQuestionNavigation('previous');
+                        }}
+                                disabled={currentQuestionIndex === 0}
+                    >
+                        Previous
+                    </button>
+                    <button
+                                className="px-4 sm:px-6 py-2.5 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-all duration-300 text-sm sm:text-base font-medium shadow-md hover:shadow-lg transform hover:-translate-y-0.5 disabled:opacity-50 disabled:cursor-not-allowed"
+                        onClick={(e) => {
+                            e.stopPropagation();
+                            handleQuestionNavigation('next');
+                        }}
+                                disabled={currentQuestionIndex === showQuestion.length - 1}
+                    >
+                        Next
+                    </button>
+                </div>
+
+                <div className="flex flex-col items-center justify-center mb-8 sm:mb-12">
+                    {currentQuestion && (
+                                <div className="mb-8 sm:mb-12 flex flex-col md:flex-row justify-center items-start gap-8 w-full">
+                            {currentQuestion.quiz.img && currentQuestion.quiz.img.length > 0 && (
+                                        <div className="w-[280px] md:w-[250px] aspect-square flex items-center justify-center bg-gray-50 rounded-xl p-4 shadow-inner">
+                                            <ImageGallery 
+                                                images={currentQuestion.quiz.img} 
+                                            />
+                                        </div>
+                            )}
+                                    <div className={`w-full md:w-1/2 ${currentQuestion.quiz.img && currentQuestion.quiz.img.length > 0 ? 'md:pl-8' : ''} flex flex-col items-center justify-center`}>
+                                        <div className="flex justify-between w-full items-center mb-6">
+                                    <h2 className="text-xl sm:text-2xl font-semibold text-gray-800">
+                                        {currentQuestion.quiz.question}
+                                    </h2>
+                                    <button
+                                        onClick={(e) => {
+                                            e.stopPropagation();
+                                            toggleBookmark(currentQuestionIndex);
+                                        }}
+                                        className="focus:outline-none transform hover:scale-110 transition-transform duration-200"
+                                    >
+                                        {currentQuestion.isBookmarked ? (
+                                            <Bookmark className="text-2xl md:text-3xl text-yellow-500 hover:text-yellow-600 transition duration-200" />
+                                        ) : (
+                                            <BookmarkBorder className="text-2xl md:text-3xl text-yellow-500 hover:text-yellow-600 transition duration-200" />
+                                        )}
+                                    </button>
+                                </div>
+                                {selectedQuestionTypes === 'mcq' ? (
+                                    <div className="flex flex-col items-center gap-4 w-full">
+                                        {shuffledChoices.map(({ choice, originalIndex }) => (
+                                            <button
+                                                key={originalIndex}
+                                                className={`px-4 sm:px-6 py-3 rounded-lg text-left w-full font-medium transition-all duration-300
+                                                    ${showQuestion[currentQuestionIndex].select === choice
+                                                        ? 'bg-blue-600 text-white shadow-md'
+                                                        : 'bg-gray-100 hover:bg-gray-200 text-gray-800'}
+                                                    ${showQuestion[currentQuestionIndex].isSubmitted ? 'cursor-not-allowed opacity-75' : ''}
+                                                `}
+                                                onClick={(e) => {
+                                                    e.stopPropagation();
+                                                    handleAnswerSelection(choice);
+                                                }}
+                                                disabled={showQuestion[currentQuestionIndex].isSubmitted}
+                                            >
+                                                {String.fromCharCode(65 + originalIndex)}. {choice}
+                                            </button>
+                                        ))}
+                                    </div>
+                                ) : selectedQuestionTypes === 'shortanswer' ? (
+                                    <div className="w-full">
+                                        <textarea
+                                            className={`w-full p-4 border-2 border-gray-200 rounded-lg focus:outline-none focus:border-blue-500 focus:ring-2 focus:ring-blue-200 transition-all duration-300
+                                                ${showQuestion[currentQuestionIndex].isSubmitted ? 'cursor-not-allowed bg-gray-50' : 'bg-white'}
+                                            `}
+                                            rows={3}
+                                            placeholder="Type your answer here..."
+                                            value={showQuestion[currentQuestionIndex].select || ''}
+                                            onClick={(e) => e.stopPropagation()}
+                                            onChange={(e) => handleShortAnswerChange(e.target.value)}
+                                            readOnly={showQuestion[currentQuestionIndex].isSubmitted}
+                                        />
+                                    </div>
+                                ) : null}
+                                <div className="flex flex-col gap-3 mt-6 w-full">
+                                    {(selectedQuestionTypes === 'mcq' || selectedQuestionTypes === 'shortanswer') && answerMode === 'each-question' && !showQuestion[currentQuestionIndex].isSubmitted && showQuestion[currentQuestionIndex].isAnswered && (
+                                        <div className="flex gap-3">
+                                            <button
+                                                className="px-4 py-2.5 bg-green-500 text-white rounded-lg hover:bg-green-600 transition-all duration-300 text-sm sm:text-base font-medium shadow-md hover:shadow-lg transform hover:-translate-y-0.5 w-1/2"
+                                                onClick={submitCurrentQuestion}
+                                            >
+                                                <CheckCircle className="mr-2" />
+                                                Submit
+                                            </button>
+                                            <button
+                                                className={`
+                                                px-4 py-2.5 bg-red-500 text-white rounded-lg hover:bg-red-600 transition-all duration-300 text-sm sm:text-base font-medium shadow-md hover:shadow-lg transform hover:-translate-y-0.5 w-1/2
+                                                ${showQuestion[currentQuestionIndex].isAnswered ? 'opacity-75 cursor-not-allowed' : ''}`}
+                                                onClick={clearAnswer}
+                                            >
+                                                <Cancel className="mr-2" />
+                                                Clear Answer
+                                            </button>
+                                        </div>
+                                    )}
+                                            {(selectedQuestionTypes === 'mcq' || selectedQuestionTypes === 'shortanswer') && answerMode === 'end-of-quiz' && showQuestion[currentQuestionIndex].isAnswered && (
+                                        <div className="flex w-full justify-end gap-3">
+                                            <button
+                                                className="px-4 py-2.5 bg-red-500 text-white rounded-lg hover:bg-red-600 transition-all duration-300 text-sm sm:text-base font-medium shadow-md hover:shadow-lg transform hover:-translate-y-0.5"
+                                                onClick={clearAnswer}
+                                            >
+                                                <Cancel className="mr-2" />
+                                                Clear Answer
+                                            </button>
+                                        </div>
+                                    )}
+                                    
+                                    {showQuestion[currentQuestionIndex].isSubmitted && (
+                                        <div className={`px-4 py-3 rounded-lg text-sm sm:text-base flex items-center font-medium
+                                            ${showQuestion[currentQuestionIndex].isCorrect === true ? 'bg-green-100 text-green-700 border border-green-200'
+                                                : showQuestion[currentQuestionIndex].isCorrect === false ? 'bg-red-100 text-red-700 border border-red-200'
+                                                    : 'bg-yellow-100 text-yellow-700 border border-yellow-200'}`}>
+                                            {showQuestion[currentQuestionIndex].isCorrect === true ?
+                                                <><CheckCircle className="mr-2" /> Correct Answer</> :
+                                                showQuestion[currentQuestionIndex].isCorrect === false ?
+                                                    <><Cancel className="mr-2" /> Wrong Answer</> :
+                                                    <><ErrorOutline className="mr-2" /> Answer Submitted</>}
+                                        </div>
+                                    )}
+                                </div>
+                                {showQuestion[currentQuestionIndex].isSubmitted && (
+                                            <div className="mt-4 p-4 bg-gray-50 rounded-lg border border-gray-200 w-full">
+                                        <p className="text-sm font-medium">Correct Answer: <span className="text-blue-600 font-semibold">{currentQuestion.quiz.correctAnswer}</span></p>
+                                    </div>
+                                )}
+                            </div>
+                        </div>
+                    )}
+                </div>
+
+                <div className="flex justify-center gap-4 mt-8">
+                    <button
+                        className="px-5 py-2.5 bg-gray-100 text-gray-700 rounded-lg hover:bg-gray-200 transition-all duration-300 text-sm sm:text-base font-medium shadow-md hover:shadow-lg transform hover:-translate-y-0.5"
+                        onClick={() => setIsQuestionTableOpen(true)}
+                    >
+                        View All Questions
+                    </button>
+                    {answerMode === 'end-of-quiz' && allQuestionsAnswered && (
+                        <button
+                            className="px-5 py-2.5 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-all duration-300 text-sm sm:text-base font-medium shadow-md hover:shadow-lg transform hover:-translate-y-0.5 disabled:opacity-50 disabled:cursor-not-allowed"
+                            onClick={handleSubmit}
+                            disabled={isSubmitting}
+                        >
+                            {isSubmitting ? "Submitting..." : "Submit All & Go to Summary"}
+                        </button>
+                    )}
+                    {answerMode === 'each-question' && allQuestionsSubmitted && (
+                        <button
+                            className="px-5 py-2.5 bg-blue-600 text-white rounded-lg hover:bg-blue-700 transition-all duration-300 text-sm sm:text-base font-medium shadow-md hover:shadow-lg transform hover:-translate-y-0.5 disabled:opacity-50 disabled:cursor-not-allowed"
+                            onClick={handleSubmit}
+                            disabled={isSubmitting}
+                        >
+                            {isSubmitting ? "Submitting..." : "Go to Summary"}
+                        </button>
+                    )}
+                        </div>
+                    </div>
+                </div>
+                </div>
+                {isQuestionTableOpen && (
+                    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50 p-4">
+                        <div className="bg-white rounded-xl shadow-2xl max-w-[95%] w-full max-h-[90vh] overflow-auto p-6 sm:p-8">
+                            <div className="flex justify-between items-center mb-6">
+                                <h2 className="text-xl font-semibold text-gray-800">All Questions</h2>
+                                <div className="flex gap-3">
+                                    <button
+                                        onClick={toggleQuestionViewMode}
+                                        className="focus:outline-none text-gray-500 hover:text-gray-700 transform hover:scale-110 transition-transform duration-200"
+                                        title="Change View"
+                                    >
+                                        {questionViewMode === 'grid' ? <ViewList className="text-xl" /> : <ViewModule className="text-xl" />}
+                                    </button>
+                                    <button
+                                        onClick={() => setIsQuestionTableOpen(false)}
+                                        className="text-gray-500 hover:text-gray-700 focus:outline-none transform hover:scale-110 transition-transform duration-200"
+                                    >
+                                        <svg className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M6 18L18 6M6 6l12 12" />
+                                        </svg>
+                                    </button>
+                                </div>
+                            </div>
+                        
+                            <div className={`grid ${questionViewMode === 'grid' ? 'grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-3 sm:gap-4' : 'flex flex-col gap-3'}`}>
+                                {showQuestion.map((q, index) => (
+                                    <button
+                                        key={index}
+                                        className={`p-3 rounded-lg text-sm sm:text-base text-left transition-all duration-300 font-medium border hover:bg-gray-100
+                                            ${index === currentQuestionIndex ? 'border-3 border-blue-400' : 'hover:bg-gray-50'}
+                                            ${q.isBookmarked ? 'border-2 border-yellow-500' : ''}
+                                            ${q.isAnswered ? 'bg-blue-100 border border-blue-200' : ''}
+                                            ${q.isSubmitted ?
+                                                q.isCorrect === true ? 'bg-green-100 text-green-800 border border-green-300'
+                                                    : q.isCorrect === false ? 'bg-red-100 text-red-800 border border-red-300'
+                                                        : 'bg-yellow-100 text-yellow-800 border border-yellow-300'
+                                                : ''}
+                                            ${questionViewMode === 'list' ? 'block w-full' : ''}
+                                            `}
+                                        onClick={() => navigateToQuestion(index)}
+                                    >
+                                        {index + 1}. {q.quiz.question.substring(0, 25)}{q.quiz.question.length > 25 ? '...' : ''}
+                                    </button>
+                                ))}
+                            </div>
+                        </div>
+                    </div>
+                )}
+                {isImageModalOpen && (
+                    <div
+                        className="fixed inset-0 bg-black bg-opacity-90 flex items-center justify-center z-50 p-4"
+                        onClick={() => setIsImageModalOpen(false)}>
+                        <div
+                            className="relative bg-white rounded-xl p-4 sm:p-6 shadow-2xl max-w-[95%] sm:max-w-4xl w-full max-h-[90vh] overflow-auto"
+                            onClick={(e) => e.stopPropagation()}>
+                            <button
+                                onClick={() => setIsImageModalOpen(false)}
+                                className="absolute top-3 right-3 text-gray-500 hover:text-gray-700 focus:outline-none transform hover:scale-110 transition-transform duration-200">
+                                <svg className="h-6 w-6" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M6 18L18 6M6 6l12 12" />
+                                </svg>
+                            </button>
+                            <div className="flex flex-col items-center justify-center h-full w-full">
+                                <div className="rounded-xl overflow-auto w-full h-full flex items-center justify-center relative bg-gray-50">
+                                <ImageGallery 
+                                    images={currentQuestion.quiz.img}
+                                    />
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                )}
         </ProtectedPage>
     );
 }
